Refactor common code for file indexer tests
